refactor(vue): normalize canvas interaction domains
