--- packages/vue/src/canvas/tool-input/index.ts
@@ -1,1 +1,0 @@
-export { handleToolMouseDown, startPanDrag } from './use'
