--- packages/vue/src/canvas/transform-input/use.ts
@@ -1,1 +1,0 @@
-export { createTransformInputActions as createCanvasTransformInput } from '#vue/canvas/transform-input/actions'
