+++ packages/vue/src/canvas/transform/input.ts
@@ -1,0 +1,1 @@
+export { createTransformInputActions as createCanvasTransformInput } from '#vue/canvas/transform/actions'
