+++ packages/vue/src/canvas/transform/pan.ts
