+++ packages/vue/src/canvas/transform/text-selection.ts
